requirements.txt
flask==2.3.2
python-dotenv==1.0.0
openai>=1.0.0
notion-client==2.0.0
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.8.5
gunicorn==21.2.0 